[#113][payment] refactor : 패키지 수정
