+++ payment/src/main/java/table/eat/now/payment/payment/infrastructure/client/ReservationClientImpl.java
@@ -3,7 +3,7 @@ package table.eat.now.payment.payment.infrastructure.client;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import table.eat.now.payment.payment.application.client.ReservationClient;
-import table.eat.now.payment.payment.application.dto.response.GetReservationInfo;
+import table.eat.now.payment.payment.application.client.dto.GetReservationInfo;
 import table.eat.now.payment.payment.infrastructure.client.feign.ReservationFeignClient;
 
 @Component
