+++ payment/src/main/java/table/eat/now/payment/payment/infrastructure/client/dto/response/CancelTossPayResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.time.LocalDateTime;
 import java.util.List;
-import table.eat.now.payment.payment.application.dto.response.CancelPgPaymentInfo;
+import table.eat.now.payment.payment.application.client.dto.CancelPgPaymentInfo;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record CancelTossPayResponse(
