+++ payment/src/main/java/table/eat/now/payment/payment/infrastructure/client/dto/response/GetReservationResponse.java
@@ -1,6 +1,6 @@
 package table.eat.now.payment.payment.infrastructure.client.dto.response;
 
-import table.eat.now.payment.payment.application.dto.response.GetReservationInfo;
+import table.eat.now.payment.payment.application.client.dto.GetReservationInfo;
 
 //Todo : internal 형식물어보기
 public record GetReservationResponse() {
